requirements.txt
absl-py==2.1.0
altair==4.2.2
asttokens==2.4.1
astunparse==1.6.3
attrs==25.2.0
backports.zoneinfo==0.2.1
blinker==1.8.2
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cycler==0.12.1
entrypoints==0.4
flatbuffers==25.2.10
fonttools==4.56.0
gast==0.4.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.70.0
h5py==3.11.0
idna==3.10
importlib-metadata==6.11.0
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.5.3
mdurl==0.1.2
narwhals==1.30.0
numpy==1.21.5
oauthlib==3.2.2
opt_einsum==3.4.0
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
pkgutil_resolve_name==1.3.10
plotly==5.11.0
protobuf==3.19.6
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydeck==0.9.1
Pympler==1.1
pyparsing==3.1.4
pytz==2025.1
pytz-deprecation-shim==0.1.0.post0
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.9.4
rpds-py==0.20.1
rsa==4.9
scikit-learn==1.0.2
scipy==1.7.3
smmap==5.0.2
streamlit==1.22.0
tenacity==8.5.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.1
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==1.0.0
tzdata==2025.1
tzlocal==4.3.1
urllib3==2.2.3
validators==0.34.0
watchdog==4.0.2
Werkzeug==3.0.6
wrapt==1.17.2
zipp==3.20.2
